src/main/java/no/kristiania/chatapp/db/UserGroupLink.java
package no.kristiania.chatapp.db;

public class UserGroupLink {
    private long id;
    private long userId;
    private long groupId;

    public void setId(long id) {
        this.id = id;
    }

    public long getId() {
        return id;
    }

    public void setUserId(long userId) {
        this.userId = userId;
    }

    public long getUserId() {
        return userId;
    }

    public void setGroupId(long groupId) {
        this.groupId = groupId;
    }

    public long getGroupId() {
        return groupId;
    }
}
